Update README.md
The first movie in the list is my favorite movie that I watch ever since I was in high school. It is my favorite because it is all about friendship and I will never get tired of watching this ever. The second is the San Andreas because I was amazed on how they created this movie, the effects and major edit got my interest on it. Last is my first ever Korean drama that I watched that's why I will never forget it.